Delete request with body
Developers in my company created a Delete endpoint that includes a request body. After some digging I learned that it was a rest-client advanced option to send a Delete with a body. So I'm proposing adding this IF or something like it to cover the edge case of a Delete with a body.